ods.py changé int en float
git-svn-id: https://cakeman-iv/svn/PI/EweeStats@96 5953bc4f-f46e-4619-b9d7-b8d55a22e21f

+++ EweeStats/ods.py
@@ -42,9 +42,9 @@ def write_ods(dataDir, analogSensors):
             dataList.append([line.rstrip() for line in di])
             
     # formatage des listes
-    timestamp = map(int, timestamp)
+    timestamp = map(float, timestamp)
     for i, elt in enumerate(dataList):
-        dataList[i] = map(int, elt)
+        dataList[i] = map(float, elt)
         
     for i, elt in enumerate(timestamp):
         sheet['A{line}'.format(line = i + 1)].set_value(elt)
